+++ src/bot/handler/__init__.py
@@ -1,0 +1,5 @@
+__all__ = [
+    'abstractHandler', 'everyoneHandler', 'groupsHandler',
+    'inlineQueryHandler', 'joinHandler', 'leaveHandler',
+    'startHandler'
+]