+++ src/bot/handler/everyoneHandler.py
@@ -5,16 +5,16 @@ from telegram.update import Update
 from bot.handler.abstractHandler import AbstractHandler
 from bot.message.messageData import MessageData
 from bot.message.replier import Replier
-from config.contents import left, not_left
-from exception.notFoundException import NotFoundException
+from config.contents import mention_failed
 from logger import Logger
 from repository.userRepository import UserRepository
+from utils.messageBuilder import MessageBuilder
 
 
-class LeaveHandler(AbstractHandler):
+class EveryoneHandler(AbstractHandler):
     bot_handler: CommandHandler
     user_repository: UserRepository
-    action: str = 'leave'
+    action: str = 'everyone'
 
     def __init__(self) -> None:
         self.bot_handler = CommandHandler(self.action, self.handle)
@@ -25,13 +25,11 @@ class LeaveHandler(AbstractHandler):
             message_data = MessageData.create_from_arguments(update, context)
         except Exception as e:
             return Replier.markdown(update, str(e))
+        
+        users = self.user_repository.get_all_for_chat(message_data.chat_id)
+        
+        if users:
+            Replier.markdown(update, MessageBuilder.mention_message(users))
+            return Logger.action(message_data, self.action)
 
-        try:
-            user = self.user_repository.get_by_id_and_chat_id(message_data.user_id, message_data.chat_id)
-            user.remove_from_chat(message_data.chat_id)
-            self.user_repository.save(user)
-
-            Replier.markdown(update, Replier.interpolate(left, message_data))
-            Logger.action(message_data, self.action)
-        except NotFoundException:
-            return Replier.markdown(update, Replier.interpolate(not_left, message_data))
+        Replier.markdown(update, mention_failed)