+++ src/bot/handler/inlineQueryHandler.py
@@ -1,21 +1,29 @@
-from bot.handler.abstractHandler import AbstractHandler
-from entity.group import Group
 from telegram import InlineQueryResultArticle
 from telegram.ext.callbackcontext import CallbackContext
-from telegram.ext.commandhandler import CommandHandler
 from telegram.ext.inlinequeryhandler import \
     InlineQueryHandler as CoreInlineQueryHandler
 from telegram.inline.inputtextmessagecontent import InputTextMessageContent
 from telegram.update import Update
 
+from bot.handler.abstractHandler import AbstractHandler
+from entity.group import Group
+from exception.actionNotAllowedException import ActionNotAllowedException
+from validator.accessValidator import AccessValidator
+
 
 class InlineQueryHandler(AbstractHandler):
-    bot_handler: CommandHandler
+    bot_handler: CoreInlineQueryHandler
 
     def __init__(self) -> None:
         self.bot_handler = CoreInlineQueryHandler(self.handle)
 
     def handle(self, update: Update, context: CallbackContext) -> None:
+        try:
+            AccessValidator.validate(str(update.effective_user.id))
+        except ActionNotAllowedException:
+            update.inline_query.answer([])
+            return
+
         group_display = update.inline_query.query or Group.default_name
         group = '' if group_display == Group.default_name else group_display
 
@@ -41,6 +49,3 @@ class InlineQueryHandler(AbstractHandler):
         ]
 
         update.inline_query.answer(results, cache_time=4800)
-
-    def get_bot_handler(self) -> CoreInlineQueryHandler:
-        return self.bot_handler