+++ src/bot/handler/startHandler.py
@@ -1,25 +1,25 @@
-from config.contents import start_text
-from logger import Logger
 from telegram.ext.callbackcontext import CallbackContext
 from telegram.ext.commandhandler import CommandHandler
 from telegram.update import Update
 
 from bot.handler.abstractHandler import AbstractHandler
 from bot.message.messageData import MessageData
+from bot.message.replier import Replier
+from config.contents import start_text
+from logger import Logger
 
 
 class StartHandler(AbstractHandler):
     bot_handler: CommandHandler
+    action: str = 'start'
 
     def __init__(self) -> None:
-        self.bot_handler = CommandHandler('start', self.handle)
+        self.bot_handler = CommandHandler(self.action, self.handle)
 
     def handle(self, update: Update, context: CallbackContext) -> None:
-        self.reply_markdown(update, start_text)
-        self.log_action(MessageData.create_from_arguments(update, context))
-
-    def get_bot_handler(self) -> CommandHandler:
-        return self.bot_handler
-
-    def log_action(self, message_data: MessageData) -> None:
-        Logger.info(f'User {message_data.username} called /start for {message_data.chat_id}')
+        try:
+            MessageData.create_from_arguments(update, context)
+        except Exception as e:
+            return Replier.markdown(update, str(e))
+        Replier.markdown(update, start_text)
+        Logger.action(MessageData.create_from_arguments(update, context), self.action)