+++ src/bot/message/messageData.py
@@ -1,19 +1,18 @@
 from __future__ import annotations
 
-import re
 from dataclasses import dataclass
 
 import names
-from entity.group import Group
-from exception.invalidArgumentException import InvalidArgumentException
 from telegram.ext.callbackcontext import CallbackContext
 from telegram.update import Update
 
+from entity.group import Group
+from validator.accessValidator import AccessValidator
 from validator.groupNameValidator import GroupNameValidator
 
 
 @dataclass
-class MessageData():
+class MessageData:
     user_id: str
     chat_id: str
     group_name: str
@@ -21,6 +20,9 @@ class MessageData():
 
     @staticmethod
     def create_from_arguments(update: Update, context: CallbackContext, include_group: bool = True) -> MessageData:
+        user_id = str(update.effective_user.id)
+        AccessValidator.validate(user_id)
+
         chat_id = str(update.effective_chat.id)
         group_name = Group.default_name
 
@@ -31,9 +33,7 @@ class MessageData():
 
             if group_name is not Group.default_name:
                 chat_id += f'~{group_name}'
-            
 
-        user_id = str(update.effective_user.id)
         username = update.effective_user.username or update.effective_user.first_name
 
         if not username: